ADCS/Sensor proc.: Send raw magnetometer readings

+++ src/ssmods/ss_ADCS_SENSORPROC/mag_io.h
@@ -19,9 +19,9 @@
 TLM_SEGMENT {
     BcTlmHeader header; // All COSMOS TLM packets must have this
 
-    double x;
-    double y;
-    double z;
+    int16_t x;
+    int16_t y;
+    int16_t z;
 } mag_segment;
 
 void magioInit();
